requirements.txt
pytorch==0.4.1
torchvision
scikit-image
cv2
tensorboardX

